[Common] Move the log.h to common.
