+++ include/common/ast/section.h
@@ -15,7 +15,7 @@
 #include "base.h"
 #include "description.h"
 #include "segment.h"
-#include "support/log.h"
+#include "common/log.h"
 #include "type.h"
 
 #include <memory>
