+++ lib/host/wasi/wasifunc.cpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 #define _DEFAULT_SOURCE
 #include "host/wasi/wasifunc.h"
+#include "common/log.h"
 #include "runtime/instance/memory.h"
 #include "support/filesystem.h"
-#include "support/log.h"
 
 #include <chrono>
 #include <csignal>
