+++ lib/interpreter/engine/engine.cpp
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 #include "common/ast/instruction.h"
+#include "common/log.h"
 #include "common/statistics.h"
 #include "common/value.h"
 #include "interpreter/interpreter.h"
 #include "support/casting.h"
-#include "support/log.h"
 #include "support/measure.h"
 #include "support/time.h"
