+++ lib/interpreter/instantiate/import.cpp
@@ -8,7 +8,7 @@
 #include "runtime/instance/memory.h"
 #include "runtime/instance/module.h"
 #include "runtime/instance/table.h"
-#include "support/log.h"
+#include "common/log.h"
 
 namespace SSVM {
 namespace Interpreter {
